Fix: dlopen 2.12 lttng-ust from fd-tracker constructor
Fix warnings:

/usr/lib64/liblttng-ust-common.so.1: undefined symbol: lttng_ust_safe_close_fd

Those are caused by the fact that liblttng-ust-common.so.1 uses dlsym to
lookup the lttng_ust_safe_close_fd symbol, but in some situations
liblttng-ust.so.0 is not loaded, because it is only dlopen from
lttng-ust-fd.c _lttng_ust_fd_ctor. liblttng-ust-common.so.1 is a common
dependency of both liblttng-ust.so.1 and liblttng-ust-fd.so.1.
Therefore, the dlopen is not performed when liblttng-ust-fd.so.1 is
preloaded.

Move the dlopen to lttng_ust_fd_tracker_init so it is performed at the
correct time.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: I292434328f512bbb284c97f6a6170edbeef7c222