Fix: ust-fork: reverse order of "after" callbacks
The "after" callbacks restore the pthread cancelstate to the prior
state, and warn on the application console when LTTNG_UST_DEBUG=1 is set
if the cancelstate restored does not match the cancelstate saved.

It is therefore important that we keep the following order:

  - begin callback (2.13)
  - begin callback (2.12)
    ...
  - after callback (2.12)
  - after callback (2.13)

to preserve the correct nesting of saved/restored pthread cancelstate.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Change-Id: Id95792654d514e0dcd09c00ff5dc0d220fb27f87